SeedKeeper support for satochip - label & label_size no longer required.
In card_bip32_import_encrypted_seed(): remove label & label_size in header.

The label and label_size were included when importing an encrypted seed into a Satochip.
With Satochip v012-0.3, this is no longer required because label & label_size are not HMACed anymore.

When importing a secret into another SeedKeeper, the label can now be modified.