Add support for 'BIP39 mnemonic v2' secrets + NFC policy
BIP39 mnemonic v2 secrets store BIP39 seeds in a compact way using the BIP39 entropy (16-32 bytes).
This allows to store the derived MasterSeed along.

NFC policy allows to enable/disable/block the NFC interface for communication.
The NFC interface is enabled by default, NFC policy can be changed using apdu command.
If NFC policy is NFC_DISABLED, NFC is not allowed but this policy can be reversed.
If NFC policy is NFC_BLOCKED, NFC is not allowed and can only be reenabled with factory reset.
The apdu command to set NFC policy must be done via the contact interface, PIN validation is required.